requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
chromadb==0.3.29
click==8.1.5
clickhouse-connect==0.6.6
coloredlogs==15.0.1
dataclasses-json==0.5.9
docx2txt==0.8
duckdb==0.8.1
exceptiongroup==1.1.2
fastapi==0.85.1
flatbuffers==23.5.26
frozenlist==1.4.0
graphlib-backport==1.0.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
langchain==0.0.232
langsmith==0.0.5
lz4==4.3.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.4
onnxruntime==1.15.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
overrides==7.3.1
packaging==23.1
pandas==2.0.3
posthog==3.0.1
protobuf==4.23.4
pulsar-client==3.2.0
pydantic==1.10.11
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.20.4
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
yarl==1.9.2
zipp==3.16.1
zstandard==0.21.0
promptlayer==0.3.3
sentence_transformers==2.2.2
